Moved regions into the commons directory
