+++ src/client/src/combineReducers.ts
@@ -3,9 +3,9 @@ import connectionStatusReducer from './connectionStatusOperations'
 import compositeStatusServiceReducer from './redux/reducers/compositeStatusServiceOperations'
 import { currencyPairReducer } from './redux/reducers/currencyPairsReducer'
 import { pricingServiceReducer } from './redux/reducers/pricingReducer'
-import { regionsReducer } from './regions/regionsOperations'
 import { analyticsReducer } from './ui/analytics/index'
 import { blotterReducer } from './ui/blotter/index'
+import { regionsReducer } from './ui/common/regions/regionsOperations'
 import footerReducer from './ui/footer/FooterOperations'
 import sidebarRegionReducer from './ui/sidebar/SidebarRegionOperations'
 import { spotTileDataReducer } from './ui/spotTile/index'
