+++ src/client/src/ui/analytics/AnalyticsContainer.tsx
@@ -6,7 +6,7 @@ import { getPositionsChartModel } from './model/positionsChartModel'
 import Analytics from './Analytics'
 import './AnalyticsStyles.scss'
 import Environment from '../../system/environment'
-import { addRegion, openWindow, regionsSettings } from '../../regions/regionsOperations'
+import { addRegion, openWindow, regionsSettings } from '../common/regions/regionsOperations'
 import { CurrencyPair } from '../../types/currencyPair'
 
 const analyticsRegionSettings = regionsSettings('Analytics', 400, 800, false)
