+++ src/client/src/ui/shell/Shell.tsx
@@ -7,7 +7,7 @@ import { BlotterContainer } from '../blotter'
 import './ShellStyles.scss'
 import '../common/styles/_base.scss'
 import '../common/styles/_fonts.scss'
-import RegionWrapper from '../../regions/RegionWrapper'
+import RegionWrapper from '../common/regions/RegionWrapper'
 import * as classnames from 'classnames'
 import TradeNotificationContainer from '../notification/TradeNotificationContainer'
 import * as PropTypes from 'prop-types'
